+++ engine/src/main/java/dev/gamekit/interfaces/Task.java
@@ -1,0 +1,7 @@
+package dev.gamekit.interfaces;
+
+/** Interface for an arbitrary task */
+public interface Task {
+  /** Abstract method implementing the task's logic */
+  void run();
+}